Adding the code for running pairwise comparisons
This works as a standalone script, but can be broken down into a frontend and set of functions. At the moment there are a lot of multidimensional arrays involved, which I'm sure could be optimised. At present the data are returned as 2d dictionaries of pairwise core and accessory distances - let me know what the more optimal data structure would be as far as the stats analysis goes!